Fixes. (#2136)
* Fixes.

* Attempt to fix contracts.

* Attempt to fix contracts.

* Fix

+++ bootstrap.sh
@@ -5,17 +5,24 @@ export CLEAN=$1
 
 # Remove all untracked files and directories.
 if [ -n "$CLEAN" ]; then
-  echo "WARNING: This will erase *all* untracked files, including hooks and submodules."
-  echo -n "Continue? [y/n] "
-  read user_input
-  if [ "$user_input" != "y" ] && [ "$user_input" != "Y" ]; then
+  if [ "$CLEAN" = "clean" ]; then
+    echo "WARNING: This will erase *all* untracked files, including hooks and submodules."
+    echo -n "Continue? [y/n] "
+    read user_input
+    if [ "$user_input" != "y" ] && [ "$user_input" != "Y" ]; then
+      exit 1
+    fi
+    rm -rf .git/hooks/*
+    git clean -fd
+    for SUBMODULE in $(git config --file .gitmodules --get-regexp path | awk '{print $2}'); do
+      rm -rf $SUBMODULE
+    done
+    git submodule update --init --recursive
+    exit 0
+  else
+    echo "Unknown command: $CLEAN"
     exit 1
   fi
-  rm .git/hooks/*
-  git clean -fd
-	for SUBMODULE in $(git config --file .gitmodules --get-regexp path | awk '{print $2}'); do
-    rm -rf $SUBMODULE
-  done
 fi
 
 git submodule update --init --recursive
